api/pyproject.toml
[tool.poetry]
name = "flask-mysql-serv"
version = "0.1.0"
description = ""
authors = ["Glicole <widyanizaz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
flask-sqlalchemy = "^3.0.0"
Flask-Session = "^0.4.0"
redis = "^4.3.4"
PyMySQL = "^1.0.2"
python-dotenv = "^0.21.0"
Flask-Cors = "^3.0.10"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
